Resolved issue with phpcs by creating a test namespace
