+++ tests/PsecioTest/Jwt/Claim/AudienceTest.php
@@ -1,14 +1,16 @@
 <?php
 
-namespace Psecio\Jwt\Claim;
+namespace PsecioTest\Jwt\Claim;
 
-class TypeTest extends \PHPUnit_Framework_TestCase
+use Psecio\Jwt\Claim\Audience;
+
+class AudienceTest extends \PHPUnit_Framework_TestCase
 {
     private $claim;
 
     public function setUp()
     {
-        $this->claim = new Type('test');
+        $this->claim = new Audience('test');
     }
 
     public function tearDown()
@@ -22,6 +24,6 @@ class TypeTest extends \PHPUnit_Framework_TestCase
     public function testGetType()
     {
         $type = $this->claim->getType();
-        $this->assertEquals('typ', $type);
+        $this->assertEquals('aud', $type);
     }
 }
