+++ tests/PsecioTest/Jwt/Claim/JwtIdTest.php
@@ -1,14 +1,16 @@
 <?php
 
-namespace Psecio\Jwt\Claim;
+namespace PsecioTest\Jwt\Claim;
 
-class CustomTest extends \PHPUnit_Framework_TestCase
+use Psecio\Jwt\Claim\JwtId;
+
+class JwtIdTest extends \PHPUnit_Framework_TestCase
 {
     private $claim;
 
     public function setUp()
     {
-        $this->claim = new Custom('test', 'tst');
+        $this->claim = new JwtId('test');
     }
 
     public function tearDown()
@@ -22,6 +24,6 @@ class CustomTest extends \PHPUnit_Framework_TestCase
     public function testGetType()
     {
         $type = $this->claim->getType();
-        $this->assertEquals('tst', $type);
+        $this->assertEquals('jti', $type);
     }
 }
